requirements.txt
pyttsx3
SpeechRecognition
pyaudio
pywhatkit
wikipedia
pyjokes
pillow
comtypes
